add window.confirm for delete-request-tech

+++ src/components/delete-request-tech/delete-request-tech.js
@@ -37,30 +37,39 @@ export default class DeleteRequestTech extends Component {
             })
     }
 
+    deleteDep(depid) {
+        if (window.confirm('Are you sure?'))
+        {
+
+        }
+    }
+
     submitHandler = (e) => {
-        e.preventDefault();
-        fetch(`http://springreact2.eba-dup8x69j.eu-central-1.elasticbeanstalk.com/api/tech/delete/${this.state.technology.id}`, {
-            method: "DELETE",
-            // mode: 'cors',
-            // headers: {
-            //     'Accept': 'application/json',
+
+        if (window.confirm('Подтвердите удаление технологии'))
+        {
+            e.preventDefault();
+            fetch(`http://springreact2.eba-dup8x69j.eu-central-1.elasticbeanstalk.com/api/tech/delete/${this.state.technology.id}`, {
+                method: "DELETE",
+                // mode: 'cors',
+                // headers: {
+                //     'Accept': 'application/json',
                 // 'Content-Type': 'application/json',
-            // },
-            // body: JSON.stringify({
-            //     name: e.target.name.value,
-            //     // email: e.target.email.value,
-            // })
-        }).then(response => response.json())
-        // .then(data => {
-        //     console.log('Data from response: ', data);
-        //     this.setState({forecasts: [...this.state.forecasts, data]})
-        // });
+                // },
+                // body: JSON.stringify({
+                //     name: e.target.name.value,
+                //     // email: e.target.email.value,
+                // })
+            }).then(response => response.json())
+            // .then(data => {
+            //     console.log('Data from response: ', data);
+            //     this.setState({forecasts: [...this.state.forecasts, data]})
+            // });
+        }
     }
 
     render() {
         const { technology } = this.state
-
-
 
         return (
             <div>
